CloudFunctions/nytimes/states/requirements.txt
# Dependencies 
SQLAlchemy == 1.3.17
pandas == 1.0.3
google-cloud-secret-manager==1.0.0
PyMySQL==0.9.3